Directory restructuring of Attachment related components
