+++ src/components/Attachment/index.js
@@ -1,2 +1,5 @@
 export { default as Attachment } from './Attachment';
+export { default as Card } from './Card';
+export { default as Audio } from './Audio';
 export { default as AttachmentActions } from './AttachmentActions';
+export { default as File } from './File';
