--- src/components/Audio/index.js
@@ -1,1 +1,0 @@
-export { default as Audio } from './Audio';
