+++ src/components/Gallery/index.js
@@ -1,1 +1,2 @@
 export { default as Gallery } from './Gallery';
+export { default as Image } from './Image';
