--- src/components/Image/index.js
@@ -1,1 +1,0 @@
-export { default as Image } from './Image';
